stash-backup: Retrieve backup archive name after SBC execution
This retrieves the file list within the backups folder in the backup
home directory and assumes that after the stash backup client was
executed, the file with the latest modification date is the latest
backup archive.

An error is emitted if the file path could not be retrieved (e.g. empty
directory) or if the filesystem entry at the given path is not a file.